requirements.txt
Django>=5.0
psycopg2-binary>=2.9
# requests
python-dotenv
# google-generativeai>=0.5.0
openai>=1.0.0
google-api-python-client
google-auth
